[FLINK-25827][task] Move and rename MultipleInputAvailabilityHelper
